+++ R/globals.R
@@ -1,1 +1,1 @@
-globalVariables(c("name", "path", "size", "modification_time", "birth_time", "active", "changed", "expires", "path"))
+globalVariables(c("id", "name", "path", "size", "modification_time", "birth_time", "active", "changed", "expires", "path"))