+++ NEWS.md
@@ -1,5 +1,6 @@
 # boxes 0.2.0 (dev version)
 
+* `box_purge`: removes expired (or all) items from (active) box (#6)
 * `pack`: arg `expires` to set an expiration date for an item (#6)
 * `pick`: get lastest item if `id = NULL` (#3)
 * `pack`: use object name as default `id`.